XFER-10666: Update the Server/User schemas and handlers
Since the most recent update of the Uluru schema we have released some
features in self service that are not yet represented in this package
for Server and User resources.

Make the schema changes using the Smithy model generated schema and
adapt the relevant code and tests to match the new features.