XFER-10648: Support SSH User keys managed externally from CFN
The User Update handler already manages SSH keys such that we only
add/remove keys CloudFormation is managing but the Read handler will
report back the actual user keys as seen by a DescribeUser request. When
keys are added outside CloudFormation this behavior would cause the CFN
to report the user has "drifted".

With this change we filter the returned collection of SSH keys to match
the expected ones under the assumption that the additional keys are
externally managed and CFN should not touch them. If the collection
returned is missing expected keys we should observe a "drift" and that
should be expected to function.