Update needs more stabilize for VPC endpoints
We learned that there are multiple cases where VPC endpoint might be in
transitional state and the Contract Test failures have consistently
pointed to the DeleteServer happening too soon adter a server update.

This change removes the shotcut in the stabilize process and forces all
code paths to go though the isVpcEndpointAvailable code path.