Make all plugin and dependencies managed in the parent pom.xml
This is a well known Maven technique to simplify managing versions in a
package with submodules. Most of the plugins and dependencies are placed
in the top level pom.xml and versions can be managed either with
dependencyManagement or properties where needed.

The RPDK plugin needs to be executed in the context of the module it is
working for so we could not consolidate that too at the top pom.xml.